git: move test helpers to internal/testutil
The two helpers in git/testutils (ParseCommitFromFixture,
ParseTagFromFixture) are used only by tests inside the git module
yet they lived under a non-internal path and were therefore part of
the module's public API.

Move the package to git/internal/testutil and drop the trailing `s`
to match the singular form already used elsewhere under internal/.
The helpers themselves are unchanged; only the import path and the
package name move.
